+++ Soomla/CCNdkBridge.cpp
@@ -31,9 +31,6 @@ USING_NS_CC;
     #include <jni.h>
     #include <string>
     #include "CCSoomlaEventDispatcher.h"
-#ifdef COCOS2D_JAVASCRIPT
-    #include "jsb/JSBinding.h"
-#endif
 
     #define CLASS_NAME "com/soomla/cocos2dx/common/NdkGlue"
 #endif
@@ -64,13 +61,9 @@ namespace soomla {
                     return;
                 }
 
-            cocos2d::CCObject *dataToPass = CCJsonHelper::getCCObjectFromJson(root);
+            cocos2d::Ref *dataToPass = CCJsonHelper::getCCObjectFromJson(root);
 
-#ifdef COCOS2D_JAVASCRIPT
-            Soomla::JSBinding::ndkCallback((cocos2d::CCDictionary *) dataToPass);
-#else
             CCSoomlaEventDispatcher::getInstance()->ndkCallback((cocos2d::CCDictionary *)dataToPass);
-#endif
 
             json_decref(root);
         }
@@ -108,6 +101,7 @@ namespace soomla {
                 const char *nativeString = t.env->GetStringUTFChars(retString, 0);
                 std::string retParamsStr(nativeString);
                 t.env->ReleaseStringUTFChars(retString, nativeString);
+                t.env->DeleteLocalRef(retString);
 
 
                 const char *jsonCharArray = retParamsStr.c_str();